Move the FS stuff into a separate file
