+++ __tests__/schemas.spec.ts
@@ -1,10 +1,10 @@
+import { Schema } from 'ajv';
 import {
     getSchemaTestCollection,
-    isValidSchema,
-    Schema,
     SchemaCollectionName,
     SchemaRecord,
-} from '../utilities';
+} from '../testutilities';
+import { isValidSchema } from '../utilities';
 
 describe('Validate examples for device to cloud', () => {
     const { schemasRecords } = getSchemaTestCollection(
@@ -15,7 +15,7 @@ describe('Validate examples for device to cloud', () => {
         // @ts-ignore
         ({ schemaName, schema, schemaTests }) => {
             test.each<Schema>(schemaTests)('%o', (example) => {
-                expect(isValidSchema(schema, example)).toEqual(true);
+                expect(isValidSchema(schema, example as any)).toEqual(true);
             });
         },
     );
@@ -30,7 +30,7 @@ describe('Validate examples for cloud to device', () => {
         // @ts-ignore
         ({ schemaName, schema, schemaTests }) => {
             test.each<Schema>(schemaTests)('%o', (example) => {
-                expect(isValidSchema(schema, example)).toEqual(true);
+                expect(isValidSchema(schema, example as any)).toEqual(true);
             });
         },
     );
@@ -45,7 +45,7 @@ describe('Validate examples for the device shadow', () => {
         // @ts-ignore
         ({ schemaName, schema, schemaTests }) => {
             test.each<Schema>(schemaTests)('%o', (example) => {
-                expect(isValidSchema(schema, example)).toEqual(true);
+                expect(isValidSchema(schema, example as any)).toEqual(true);
             });
         },
     );
