+++ index.ts
@@ -1,2 +1,2 @@
-export * from './utilities';
+export { isValidSchema, getValidation } from './utilities';
 export * from './schemas';
