+++ package.json
@@ -11,8 +11,7 @@
     "tslint": "tslint --project ./tsconfig.json",
     "prettier": "prettier --write \"./__tests__/**/*.ts\"",
     "prettier:lint": "prettier --list-different \"./__tests__/**/*.ts\"",
-    "build": "npx rollup -c rollup.config.js",
-    "postinstall": "npm run build"
+    "build": "rollup -c rollup.config.js"
   },
   "repository": {
     "type": "git",
@@ -41,7 +40,7 @@
     "@rollup/plugin-node-resolve": "^13.0.6",
     "@rollup/plugin-typescript": "^8.3.0",
     "@types/glob": "^7.1.3",
-    "@types/jest": "^27.0.1",
+    "@types/jest": "^27.0.2",
     "@types/node": "^16.0.0",
     "ajv": "^8.6.1",
     "glob": "^7.1.7",
@@ -49,6 +48,7 @@
     "jest": "^27.0.6",
     "prettier": "^2.3.2",
     "rollup": "^2.59.0",
+    "rollup-plugin-polyfill-node": "^0.7.0",
     "ts-jest": "^27.0.3",
     "tslint": "^5.20.1",
     "typescript": "^4.3.5"
